Reducing split to make more readable
